Code refactoring: Combine all services into a single citation_service.py as methods. Delete old service files. Rename tests. Yet to be determined if repositories and tests should also be combined into a single file or not.
