+++ src/tests/test_inproceedings_validation.py
@@ -1,10 +1,10 @@
 import pytest
-from services.inproceedings_service import validate_inproceedings, UserInputError
+from services.citation_service import validate_inproceedings, UserInputError
 from unittest.mock import patch
 
 def test_validate_inproceedings_valid():
     """Test that a valid inproceedings passes validation and is created."""
-    with patch("services.inproceedings_service.inproceedings_repository.create_inproceedings") as mock_create:
+    with patch("services.citation_service.inproceedings_repository.create_inproceedings") as mock_create:
         mock_create.return_value = 1
 
         result = validate_inproceedings(
@@ -28,7 +28,7 @@ def test_validate_inproceedings_valid():
 
 def test_validate_inproceedings_missing_optional_fields():
     """Test that an inproceedings can be created without optional fields."""
-    with patch("services.inproceedings_service.inproceedings_repository.create_inproceedings") as mock_create:
+    with patch("services.citation_service.inproceedings_repository.create_inproceedings") as mock_create:
         mock_create.return_value = 1
 
         result = validate_inproceedings(
